Adds a new control for showing all courses the student is enrolled in
Additionally theres a new function in locallib to match labels against values, incase the samples returned
for all courses have differing amounts of results.
E.g. course 1 with 14 results vs course 2 with 13 results.
This is needed as chartjs requires the number of data points to match the number of labels